+++ wroprg/src/hardware/__init__.py
@@ -1,7 +1,8 @@
 """This module initializes the hardware package by importing necessary classes"""
 from .hardware_interface import HardwareInterface
 from .validator import RobotValidator
-from .hardware_interface import RobotState
+from .robotstate import RobotState
+from .orientation import OrientationEstimator
 
 # We want to expose the HardwareInterface and RobotValidator classes
 # to be used in other modules, so we define them in __all__.
